Remove explicit use of pydicom in dicom_parser, dicompyler keeps the pydicom.FileDataset
