+++ dvha/db/dicom_parser.py
@@ -8,7 +8,8 @@
 #    See the file LICENSE included with this distribution, also
 #    available at https://github.com/cutright/DVH-Analytics
 
-from dicompylercore import dicomparser, dvhcalc
+from dicompylercore import dvhcalc
+from dicompylercore.dicomparser import DicomParser as dicompylerParser
 from datetime import datetime
 from dateutil.relativedelta import relativedelta  # python-dateutil
 from dateutil.parser import parse as date_parser
@@ -16,14 +17,12 @@ import numpy as np
 from os.path import basename, join
 from pubsub import pub
 from dvha.tools.roi_name_manager import clean_name, DatabaseROIs
-from dvha.tools.utilities import change_angle_origin, calc_stats, is_date
+from dvha.tools.utilities import change_angle_origin, calc_stats, is_date, validate_transfer_syntax_uid
 from dvha.tools.roi_formatter import dicompyler_roi_coord_to_db_string, get_planes_from_string
 from dvha.tools import roi_geometry as roi_calc
 from dvha.tools.mlc_analyzer import Beam as mlca
 from dvha.db.sql_connector import DVH_SQL
 from dvha.paths import IMPORT_SETTINGS_PATH, parse_settings_file
-import pydicom
-from pydicom.uid import ImplicitVRLittleEndian
 
 
 class DICOM_Parser:
@@ -65,21 +64,20 @@ class DICOM_Parser:
         self.dicompyler_rt_plan = None
         self.dicompyler_rt_structures = None
 
-        # TODO: Remove the need for both pydicom and dicompyler DICOM parsers
         if self.plan_file:
-            self.rt_data['plan'] = pydicom.read_file(self.plan_file, force=True)
-            self.dicompyler_data['plan'] = dicomparser.DicomParser(self.plan_file)
+            self.dicompyler_data['plan'] = dicompylerParser(self.plan_file)
+            self.rt_data['plan'] = self.dicompyler_data['plan'].ds
             self.dicompyler_rt_plan = self.dicompyler_data['plan'].GetPlan()
         if self.structure_file:
-            self.rt_data['structure'] = pydicom.read_file(self.structure_file, force=True)
-            self.dicompyler_data['structure'] = dicomparser.DicomParser(self.structure_file)
+            self.dicompyler_data['structure'] = dicompylerParser(self.structure_file)
+            self.rt_data['structure'] = self.dicompyler_data['structure'].ds
             self.dicompyler_rt_structures = self.dicompyler_data['structure'].GetStructures()
         if self.dose_file:
-            self.rt_data['dose'] = pydicom.read_file(self.dose_file, force=True)
             if self.dose_sum_file is None:
-                self.dicompyler_data['dose'] = dicomparser.DicomParser(self.dose_file).ds
+                self.dicompyler_data['dose'] = dicompylerParser(self.dose_file)
             else:
-                self.dicompyler_data['dose'] = dicomparser.DicomParser(self.dose_sum_file).ds
+                self.dicompyler_data['dose'] = dicompylerParser(self.dose_sum_file)
+            self.rt_data['dose'] = self.dicompyler_data['dose'].ds
 
         # These properties are not inherently stored in Pinnacle DICOM files, but can be extracted from dummy ROI
         # names automatically generated by the Pinnacle Script provided by DVH Analytics
@@ -397,9 +395,8 @@ class DICOM_Parser:
         except AttributeError as e:
             # print(str(e), 'for MRN: %s' % self.mrn)
             # print('Applying validate_transfer_syntax_uid() due to missing data in user provided DICOM')
-            dose_file = self.dose_sum_file if self.dose_sum_file else self.rt_data['dose']
-            dose = self.validate_transfer_syntax_uid(dose_file)
-            structure = self.validate_transfer_syntax_uid(self.rt_data['structure'])
+            dose = validate_transfer_syntax_uid(self.rt_data['dose'])
+            structure = validate_transfer_syntax_uid(self.rt_data['structure'])
             dvh = dvhcalc.get_dvh(structure, dose, dvh_index,
                                   callback=self.send_dvh_progress)
 
@@ -934,18 +931,6 @@ class DICOM_Parser:
                 if rad_type in beam.RadiationType:
                     return True
         return False
-
-    @staticmethod
-    def validate_transfer_syntax_uid(data_set):
-        meta = pydicom.Dataset()
-        meta.ImplementationClassUID = pydicom.uid.generate_uid()
-        meta.TransferSyntaxUID = ImplicitVRLittleEndian
-        new_data_set = pydicom.FileDataset(filename_or_obj=None, dataset=data_set, is_little_endian=True,
-                                           file_meta=meta)
-        new_data_set.is_little_endian = True
-        new_data_set.is_implicit_VR = True
-
-        return new_data_set
 
     @property
     def pre_import_data(self):
